Add files via upload
The dynamical state of galaxy groups at intermediate redshifts can provide information about the growth of structure in the universe. We examine three goodness-of-fit tests, the Anderson–Darling (A–D), Kolmogorov, and χ2 tests, in order to determine which statistical tool is best able to distinguish between groups that are relaxed and those that are dynamically complex. We perform Monte Carlo simulations of these three tests and show that the χ2 test is profoundly unreliable for groups with fewer than 30 members. Power studies of the Kolmogorov and A–D tests are conducted to test their robustness for various sample sizes, where we use the Skew Normal Distribution. We then apply these non-parametric tests to a sample from the 2MASS catalog. We applied these well-known normality tests and more tests, such as Lilliefors, D’Agostino, Shapiro Wilk, Cramer-von Misses and Wilcoxon to the velocity and magnitude distributions of these systems to distinguish Gaussian and non-Gaussian clusters. Also we study the Rayleigh distribution. We analyzed galaxy samples from the R package “MASS”, the Astro-statistics dataset (NASA astrophysics data system (A.M. Garcia 1992), COMBO-17), where we show important histograms for the normality of the group galaxies and uniformity (Rayleigh and Wilcoxon tests) for the morphological type of galaxies (Elliptical, Spiral, Lenticular). Furthermore we use a sample from the Canadian Network for Observational Cosmology (CNOC 2) Field Galaxy Redshift Survey, where we approve some figures from the papers Hou.et al. Finally we compute velocity dispersion profiles(VDPs) for all groups of the CNOC 2 sample with more than 20 members and compare the overall features of the Gaussian and non-Gaussian groups, finding that the VDPs of the non-Gaussian groups are distinct and dynamically more complex from those classified as Gaussian. An the end, we compared the results of the velocities, the masses and the magnitudes with Gaussian and Non-Gaussian galaxies. Our achievement is to apply non-parametric statistical tools for group galaxies we the aim of distinguishing Normal and Non-Normal Groups.